Move packet.rs -> packets/mod.rs, define mods for all packets
