+++ src/packets/publish.rs
@@ -3,13 +3,3 @@
 //   License, v. 2.0. If a copy of the MPL was not distributed with this
 //   file, You can obtain one at http://mozilla.org/MPL/2.0/.
 //
-
-mod client;
-mod client_identifier;
-mod codecs;
-mod error;
-mod keep_alive;
-mod packets;
-mod properties;
-mod string;
-mod transport;
